[addon] rename AudioEncoder function get_addon to SetAudioEncoderFuncTable
Is to prevent conflicts if more as one add-on type is used.
Also is it no more needed to export them and only be present to make for
add-on more easy to set function table on ADDON_CreateInstance call.

On final it becomes removed complete and be done from dev kit C++ headers.